split up currency tokens from contract addresses
