+++ src/constants/ethContractAddresses.ts
@@ -13,24 +13,3 @@ export const usdcTokenAddress = process.env.REACT_APP_USDC_TOKEN_ADDRESS
 export const uniswapEthDpiLpTokenAddress =
   process.env.REACT_APP_UNISWAP_ETH_DPI_LP_TOKEN_ADDRESS
 export const uniswapRouterAddress = process.env.REACT_APP_UNISWAP_ROUTER_ADDRESS
-
-export const currencyTokens = [
-  {
-    name: 'Ether',
-    label: 'ETH',
-    id: 'wrapped_eth',
-    address: undefined,
-  },
-  {
-    name: 'Dai',
-    label: 'DAI',
-    id: 'mcd',
-    address: daiTokenAddress,
-  },
-  {
-    name: 'USD Coin',
-    label: 'USDC',
-    id: 'usdc',
-    address: usdcTokenAddress,
-  },
-]
